Refactor class accessor checks
